ml_service/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
joblib==1.3.1
requests==2.31.0
